(finally) completed line autoindentation
*jeeze* talk about a pain in the rear. I had to fix both the set and get cursor system for this, the biggest struggle was how HTML won't insert the <brs> right if it is the first br for the cell (if that makes sense). The silly solution to that problem is adding a break to the end of the source in those cases.
```julia
    if ~(contains(cell.source, "\n") || contains(cell.source, "<br>"))
        interior = component[:children]["cellinterior$(cell.id)"]
        inp = interior[:children]["cellinput$(cell.id)"]
        maincell::Component{:div} = inp[:children]["cell$(cell.id)"]
        maincell[:text] = maincell[:text] * "<br>"
    end
```

Ugh. So, when it gets a <br> and isn't ready for one -- of all the things to do, it blurs. This would make the **first** enter NEVER WORK and unfocus the cell, very annoying, but after that the autocomplete would be normal. This was a nightmare, honestly! Hehe. I am happy with its state here, even if it isn't exactly perfect...  The last remaining feature to create will be the actual autocomplete insert functionality, which I think will actually be a lot easier. No idea as to whether or not this is unicode safe. The first iteration may not be.